Break CoordSeqIter into its own module
